refactor: app structure
