+++ src/constants/index.js
@@ -4,6 +4,12 @@ export const DEFAULT_USERNAME = 'Anonymous';
 export const INVALID_INPUT = 'Invalid input';
 export const OPERATION_FAILED = 'Operation failed';
 
-export const COMMANDS = {
-    EXIT: '.exit',
+export const EXIT_COMMAND = '.exit';
+
+export const OS_COMMAND_ARGUMENTS = {
+    EOL: '--EOL',
+    CPUS: '--cpus',
+    HOMEDIR: '--homedir',
+    USERNAME: '--username',
+    ARCHITECTURE: '--architecture',
 };
