+++ src/modules/commands/exit/index.js
@@ -1,7 +1,7 @@
-import { printMessage } from '../../utils/printMessage.js';
-import { user } from '../userInfo.js';
+import { printMessage } from '../../../utils/printMessage.js';
+import { user } from '../../userInfo.js';
 
-import { INVALID_INPUT } from '../../constants/index.js';
+import { INVALID_INPUT } from '../../../constants/index.js';
 
 export const exit = (args) => {
     if (args && args.length) return printMessage(INVALID_INPUT);
