+++ src/modules/commands/operationManager.js
@@ -1,0 +1,7 @@
+import { exit } from './exit/index.js';
+
+import { EXIT_COMMAND } from '../../constants/index.js'
+
+export const operationManager = {
+    [EXIT_COMMAND]: exit
+};
